requirements.txt
crewai==0.105.0
crewai[tools]
langchain_openai
langchain_community
python-dotenv
matplotlib
networkx
mdutils
beautifulsoup4
scholarly
pandas